requirements.txt
aiocoap==0.3
asyncio==3.4.3
paho-mqtt==1.5.1
pkg-resources==0.0.0
